fix: resolve server action fetchServerAction error - convert server actions to API routes for better client-server communication
